fraction

Module generic

Integer generic traits and operations

These traits and functions are for use in generic algorithms that work indifferently to particular types, relying on some common traits.

Traits

GenericInteger

Methods common to all integer types that could be used generically in abstract algorithms

Functions

read_generic_integer

Builds integer of type T from another integer of type F in a generic way. Guarantees to only return positive values.