Struct fraction::prelude::GenericDecimal

pub struct GenericDecimal<T, P>(/* private fields */)
where
    T: Clone + Integer,
    P: Copy + Integer + Into<usize>;

Decimal type implementation

T is the type for data P is the type for precision

Uses GenericFraction internally to represent the data. Precision is being used for representation purposes only. Calculations do not use precision, but comparison does.

Examples

use fraction::GenericDecimal;

type Decimal = GenericDecimal<u64, u8>;

let d1 = Decimal::from(12);
let d2 = Decimal::from(0.5);

let mul = d1 * d2;
let div = d1 / d2;
let add = d1 + d2;

assert_eq!(mul, 6.into());
assert_eq!(div, Decimal::from("24.00"));
assert_eq!(add, Decimal::from(12.5));

Implementations

impl<T, P> GenericDecimal<T, P>
where T: Clone + GenericInteger, P: Copy + GenericInteger + Into<usize>,

pub const fn sign(&self) -> Option<Sign>
where T: CheckedAdd + CheckedMul + CheckedSub,

Returns Some(Sign) of the decimal, or None if NaN is the value

pub fn set_precision(self, precision: P) -> Self

Sets representational precision for the Decimal The precision is only used for comparison and representation, not for calculations.

This is suggested method for you to utilise if you know what precision you want to work with.

Examples

use fraction::GenericDecimal;

type D = GenericDecimal<u32, u8>;

let first = D::from("0.004")  // initial precision is 4
            .set_precision(2);  // but we want to work with 2
let second = D::from("0.006").set_precision(2);

// Even though "first" and "second" both have precision 2
// the actual calculations are still performed with their
// exact initial precision
assert_eq!(first + second, D::from("0.01"));

// The comparison, on the other hand, takes the precision into account
// so the actual compared numbers will be calculated up the highest
// precision of both operands
assert_ne!(  // compares "0.010" with "0.011"
    D::from("0.01"),  // has precision 2
    D::from("0.011")  // has precision 3
);

assert_eq!(  // compares "0.01" with "0.01"
    D::from("0.01").set_precision(2),
    D::from("0.011").set_precision(2)
);

pub const fn get_precision(&self) -> P

Returns the current representational precision for the Decimal

pub fn calc_precision(self, max_precision: Option<P>) -> Self
where T: CheckedMul + DivAssign + MulAssign + SubAssign + ToPrimitive + GenericInteger, P: Bounded + CheckedAdd,

Try to recalculate the representational precision depending on the internal Fraction, which is the actual value.

Performs the actual division until the exact decimal value is calculated, the precision type (P) capacity is reached (e.g. 255 for u8) or max_precision is reached, if it is given.

WARNING

You only need this method if you want to find the max available precision for the current decimal value. However, keep in mind that irrational values (such as 1/3) do not have finite precision, so if this method returns P::MAX (or max_precision), most likely you have an irrational value. Be careful with max numbers for usize - that can take very long time to compute (more than a minute)

pub fn nan() -> Self

pub fn infinity() -> Self

pub fn neg_infinity() -> Self

pub fn neg_zero() -> Self

pub fn min_positive_value() -> Self
where T: Bounded, P: Bounded,

pub const fn is_nan(&self) -> bool

pub const fn is_infinite(&self) -> bool

pub const fn is_finite(&self) -> bool

pub fn is_normal(&self) -> bool

pub fn classify(&self) -> FpCategory

pub fn floor(&self) -> Self

pub fn ceil(&self) -> Self

pub fn round(&self) -> Self

pub fn trunc(&self) -> Self

pub fn fract(&self) -> Self

pub fn abs(&self) -> Self

pub fn signum(&self) -> Self

pub const fn is_sign_positive(&self) -> bool

pub const fn is_sign_negative(&self) -> bool

pub fn mul_add(&self, a: Self, b: Self) -> Self

pub fn recip(&self) -> Self

pub fn map( self, fun: impl FnOnce(GenericFraction<T>) -> GenericFraction<T>, ) -> Self

pub fn map_mut(&mut self, fun: impl FnOnce(&mut GenericFraction<T>))

pub fn map_ref( &self, fun: impl FnOnce(&GenericFraction<T>) -> GenericFraction<T>, ) -> Self

pub fn apply_ref<R>(&self, fun: impl FnOnce(&GenericFraction<T>, P) -> R) -> R

๐Deprecated: Use match decimal {GenericDecimal(fraction, precision) => ... }

pub fn from_fraction(fraction: GenericFraction<T>) -> Self
where T: GenericInteger + ToPrimitive, P: Bounded + CheckedAdd,

Convert from a GenericFraction

Automatically calculates precision, so for โbadโ numbers may take a lot of CPU cycles, especially if precision represented by big types (e.g. usize)

Examples

use fraction::{Fraction, Decimal};

let from_fraction = Decimal::from_fraction(Fraction::new(1u64, 3u64));
let from_division = Decimal::from(1) / Decimal::from(3);

let d1 = Decimal::from(4) / from_fraction;
let d2 = Decimal::from(4) / from_division;

assert_eq!(d1, d2);
assert_eq!(d1, Decimal::from(12));