Enum fraction::dynaint::DynaInt

pub enum DynaInt<T, G>
where
    T: Copy + GenericInteger + Into<G> + TryToConvertFrom<G> + From<u8>,
    G: Clone + GenericInteger,
{
    S(T),
    __H(G),
}

The wrapper implementation

Keeps data within S (small) whenever possible, performing checked arithmetic and moving onto __H (huge) when overflows happen. Every math operation on __H performs a read of the resulting number with TryToConvertFrom::try_to_convert_from. Every math operation on S is checked for overflows.

Variants

S(T)

Represents the small type, implementing Copy and allocated on stack. The wrapper tries to reduce contained values to this type whenever possible

__H(G)

Represents the huge type, implementing Clone. To be used when values overflow T

Implementations

impl<T, G> DynaInt<T, G>
where T: Copy + GenericInteger + Into<G> + TryToConvertFrom<G> + From<u8>, G: Clone + GenericInteger,

pub fn unpack(self) -> Result<T, G>

Unpacks the value

Utilises Result::Ok for S(small) numbers and Result::Err for __H(huge) ones

Examples

use fraction::dynaint::DynaInt;
type D = DynaInt<u8, u16>;

assert_eq!(Ok(1u8), D::from(1u8).unpack());
assert_eq!(Err(256u16), D::from(256u16).unpack());