Module fraction::prelude

Predefines some types for the most common use cases

You should consider this module as a list of shortcuts, and not as the list of only available types. The actual workhorses are a part of the Public API and you are encouraged to use them straightforwardly whenever you may feel necessary.

Long story short, you may compose your own types with these:

GenericFraction for fractions

GenericDecimal for decimals

DynaInt integers on stack, but dynamically growing into heap when necessary

Re-exports

pub use super::dynaint::DynaInt;

Structs

BigInt

A big signed integer type.

BigUint

A big unsigned integer type.

GenericDecimal

Decimal type implementation

Enums

GenericFraction

Generic implementation of the fraction type

Type Definitions

BigDecimal

Heap allocated BigUint for numerics and usize for precision

BigFraction

Fraction consisting from two BigUint numbers

Decimal

Basic Decimal based on 2 u64 numbers and one u8 for precision. Able to keep up to 19 digits in the number (including both sides across the floating point).

DynaDecimal

Stack allocated, but dynamically growing into heap if necessary

DynaFraction

Stack allocated, but dynamically growing into heap if necessary

Fraction

Fraction consisting from two u64 numbers