fraction

Module prelude

Predefines some types for the most common use cases

You should consider this module as a list of shortcuts, and not as the list of only available types. The actual workhorses are a part of the Public API and you are encouraged to use them straightforwardly whenever you may feel necessary.

Long story short, you may compose your own types with these:

GenericFraction for fractions

GenericDecimal for decimals

DynaInt integers on stack, but dynamically growing into heap when necessary

Structs

BigInt

A big signed integer type.

BigUint

A big unsigned integer type.

Enums

GenericFraction

Generic implementation of the fraction type

Type Aliases

BigFraction

Fraction consisting from two BigUint numbers

Fraction

Fraction consisting from two u64 numbers